% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_piat}
\alias{demo_piat}
\title{Demo PIAT}
\usage{
demo_piat(num_items = 10L, take_training = TRUE,
  feedback = psychTestRCAT::cat.feedback.graph("PIAT"),
  admin_password = "demo", researcher_email = NULL, ...)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{take_training}{(Logical scalar) Whether to take the training phase.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test. Defaults to a graph-based feedback page.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{...}{Further arguments to be passed to \code{\link{piat}()}.}
}
\description{
This function launches a demo for the PIAT.
}
