% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{piat.feedback.simple_score}
\alias{piat.feedback.simple_score}
\title{PIAT feedback (simple score)}
\usage{
piat.feedback.simple_score(dict = piat::piat_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant's score is reported at the end of the test.
}
