% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piat.R
\name{piat}
\alias{piat}
\title{PIAT}
\usage{
piat(num_items = 25L, take_training = TRUE, label = "PIAT",
  feedback = piat.feedback.no_score(dict = dict),
  media_dir = "http://media.gold-msi.org/test_materials/PIAT/1-0-1/mp4",
  dict = piat::piat_dict)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{take_training}{(Logical scalar) Whether to include the training phase.}

\item{label}{(Character scalar) Label to give the PIAT results in the output file.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test. By default no feedback is given.}

\item{media_dir}{(Character scalar) File path to the directory
hosting the test's media files (typically a publicly accessible web directory).}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines an pitch imagery ability test (PIAT)
module for incorporation into a psychTestR timeline.
Use this function if you want to include the PIAT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the PIAT, consider using \code{\link{demo_piat}()}.
For a standalone implementation of the PIAT,
consider using \code{\link{standalone_piat}()}.
}
