% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-piat.R
\name{standalone_piat}
\alias{standalone_piat}
\title{Standalone PIAT}
\usage{
standalone_piat(
  title = "Pitch imagery ability test",
  admin_password = "replace-with-secure-password",
  researcher_email = NULL,
  languages = piat_languages(),
  dict = piat::piat_dict,
  prepend_interleaving_trial_page = NULL,
  append_interleaving_trial_page = NULL,
  get_p_id = TRUE,
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}) and German (\code{"DE"}).}

\item{dict}{The dictionary to use.}

\item{prepend_interleaving_trial_page}{An optional interleaving page to come before the main trial page.}

\item{append_interleaving_trial_page}{An optional interleaving page to come after the main trial page.}

\item{get_p_id}{Should the p_id be colelcted from the user?}

\item{...}{Further arguments to be passed to \code{\link{piat}()}.}

\item{post_training_tl}{An optional timeline to displayed post-training.}
}
\description{
This function launches a standalone testing session for the PIAT.
This can be used for data collection, either in the laboratory or online.
}
