% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{piat.feedback.no_score}
\alias{piat.feedback.no_score}
\title{PIAT feedback (no score)}
\usage{
piat.feedback.no_score(dict = piat::piat_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given no feedback at the end of the test.
}
