% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-piat.R
\name{standalone_piat}
\alias{standalone_piat}
\title{Standalone PIAT}
\usage{
standalone_piat(
  title = "Pitch imagery ability test",
  admin_password = "replace-with-secure-password",
  researcher_email = NULL,
  languages = piat_languages(),
  dict = piat::piat_dict,
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}) and German (\code{"DE"}).}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{piat}()}.}
}
\description{
This function launches a standalone testing session for the PIAT.
This can be used for data collection, either in the laboratory or online.
}
